job1-api-to-gcs/requirements.txt
rx                   # Thư viện ReactiveX (RxPY)
requests             # Để thực hiện HTTP call đến MockAPI
google-cloud-storage # Để upload file lên GCS